refactor: separate Moments from ProfileGallery
